refactor: enhance ZkProof service, to be only one singleton class to gen RLN or Semaphore proofs
