--- src/background/services/zkProof/ZkProof.ts
@@ -1,5 +1,0 @@
-import ZkIdentityDecorater from "@src/background/identityDecorater";
-
-export abstract class ZkProof<Request, Return> {
-  protected abstract genProof(identity: ZkIdentityDecorater, request: Request): Promise<Return>;
-}
