--- src/contentScripts/proof/index.ts
@@ -1,6 +1,0 @@
-/**
- * Proof generation is running on content-script side.
- * It's because it's not possible to send data back to nested web worker from service worker.
- */
-export * from "./rln";
-export * from "./semaphore";
