+++ src/types/proof/RLNProof.ts
@@ -5,3 +5,8 @@ export type { RLNFullProof, VerificationKey } from "rlnjs";
 export interface RLNProofRequest extends SemaphoreProofRequest {
   rlnIdentifier: string;
 }
+
+export interface IRlnGenerateArgs {
+  identity: string;
+  payload: RLNProofRequest;
+}
